+++ src/renderer/uikit/Notification/AlertItem.ts
@@ -4,6 +4,7 @@ export interface AlertData {
     message: string;
     type: TMessageType;
     key: number;
+    createdAt: number;
     onClose: (value?: unknown) => void;
 }